+++ include/stdexec/__detail/__basic_sender.hpp
@@ -24,8 +24,6 @@
 #include "__tuple.hpp"
 #include "__type_traits.hpp"
 
-#include "../concepts.hpp"
-
 #include <utility> // for tuple_size/tuple_element
 #include <cstddef>
 #include <type_traits>
