+++ include/stdexec/__detail/__completion_signatures.hpp
@@ -185,4 +185,4 @@ namespace stdexec {
 
   template <class _Sender, class _Env>
   using __completion_signatures_of_t = __call_result_t<get_completion_signatures_t, _Sender, _Env>;
-}
+} // namespace stdexec
