+++ include/stdexec/__detail/__meta.hpp
@@ -776,15 +776,6 @@ namespace stdexec {
   template <class _Fn>
   __conv(_Fn) -> __conv<_Fn>;
 
-  // Implemented as a class instead of a free function
-  // because of a bizarre nvc++ compiler bug:
-  struct __cref_fn {
-    template <class _Ty>
-    auto operator()(const _Ty&) -> const _Ty&;
-  };
-  template <class _Ty>
-  using __cref_t = decltype(__cref_fn{}(__declval<_Ty>()));
-
   template <class, class, class, class>
   struct __mzip_with2_;
